refactor Dashboard
